Refactor: QueryHelper class in it's own file
