+++ ols/src/query_helpers/query_helper.py
@@ -1,0 +1,29 @@
+"""Base class for query helpers."""
+
+import logging
+from typing import Optional
+
+from ols.src.llms.llm_loader import LLMLoader
+from ols.utils import config
+
+logger = logging.getLogger(__name__)
+
+
+class QueryHelper:
+    """Base class for query helpers."""
+
+    def __init__(
+        self,
+        provider: Optional[str] = None,
+        model: Optional[str] = None,
+        llm_params: Optional[dict] = None,
+        llm_loader: Optional[LLMLoader] = None,
+    ) -> None:
+        """Initialize query helper."""
+        # NOTE: As signature of this method is evaluated before the config,
+        # is loaded, we cannot use the config directly as defaults and we
+        # need to use those values in the init evaluation.
+        self.provider = provider or config.ols_config.default_provider
+        self.model = model or config.ols_config.default_model
+        self.llm_params = llm_params or {}
+        self.llm_loader = llm_loader or LLMLoader
